+++ src/gb/gbGfx.cpp
@@ -1,6 +1,6 @@
 #include <memory.h>
 
-#include "../Util.h"
+#include "../System.h"
 #include "gbGlobals.h"
 #include "gbSGB.h"
 
@@ -53,7 +53,7 @@ void gbRenderLine()
         bank1 = &gbVram[0x2000];
     } else {
         bank0 = &gbMemory[0x8000];
-        bank1 = NULL;
+        bank1 = nullptr;
     }
 
     int tile_map = 0x1800;
@@ -90,7 +90,7 @@ void gbRenderLine()
     int tile_map_address = tile_map_line_y + tx;
 
     uint8_t attrs = 0;
-    if (bank1 != NULL)
+    if (bank1 != nullptr)
         attrs = bank1[tile_map_address];
 
     uint8_t tile = bank0[tile_map_address];
@@ -404,7 +404,7 @@ void gbDrawSpriteTile(int tile, int x, int y, int t, int flags,
         bank1 = &gbVram[0x2000];
     } else {
         bank0 = &gbMemory[0x8000];
-        bank1 = NULL;
+        bank1 = nullptr;
     }
 
     int SpritesTicks = gbSpritesTicks[x + 8] * (gbSpeed ? 2 : 4);
