+++ src/gb/gbSound.cpp
@@ -2,7 +2,7 @@
 
 #include <cstring>
 
-#include "../Util.h"
+#include "core/base/file_util.h"
 #include "../gba/Sound.h"
 #include "gb.h"
 #include "gbGlobals.h"
