+++ src/gba/Sound.cpp
@@ -3,7 +3,7 @@
 #include <array>
 #include <cstring>
 
-#include "../Util.h"
+#include "core/base/file_util.h"
 #include "../common/Port.h"
 #include "GBA.h"
 #include "Globals.h"
